More warnings fixed.

+++ src/Bpp/Seq/AlphabetIndex/AAChargeIndex.h
@@ -133,12 +133,12 @@ public:
   double getIndex(int state) const throw (BadIntException)
   {
     if (state < 0 || state > 19) throw BadIntException(state, "AAChargeIndex::getIndex(). Invalid state.", &AlphabetTools::PROTEIN_ALPHABET);
-    return charge_[state];
+    return charge_[static_cast<size_t>(state)];
   }
 
   double getIndex(const std::string& state) const throw (BadCharException)
   {
-    return charge_[AlphabetTools::PROTEIN_ALPHABET.charToInt(state)];
+    return charge_[static_cast<size_t>(AlphabetTools::PROTEIN_ALPHABET.charToInt(state))];
   }
 
   std::vector<double>* getIndexVector() const { return new std::vector<double>(charge_); }
