+++ src/Bpp/Seq/AlphabetIndex/AAChenGuHuangHydrophobicityIndex.h
@@ -104,12 +104,12 @@ public:
   double getIndex(int state) const throw (BadIntException)
   {
     if (state < 0 || state > 19) throw BadIntException(state, "AAChenGuHuangHydrophobicityIndex::getIndex(). Invalid state.", &AlphabetTools::PROTEIN_ALPHABET);
-    return hydrophobicity_[state];
+    return hydrophobicity_[static_cast<size_t>(state)];
   }
 
   double getIndex(const std::string& state) const throw (BadCharException)
   {
-    return hydrophobicity_[AlphabetTools::PROTEIN_ALPHABET.charToInt(state)];
+    return hydrophobicity_[static_cast<size_t>(AlphabetTools::PROTEIN_ALPHABET.charToInt(state))];
   }
 
   std::vector<double>* getIndexVector() const { return new std::vector<double>(hydrophobicity_); }
